requirements.txt
python-telegram-bot
transformers
torch
torchvision
Pillow
requests
python-dotenv 